chore(settings): refactor settings view
