+++ src/containers/SideBar/components/Settings/Markups/MoneroAddressMarkup/MoneroAddressEditor.tsx
