+++ src/containers/SideBar/components/Settings/Markups/MoneroAddressMarkup/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MoneroAddressMarkup';
