+++ src/containers/SideBar/components/Settings/Markups/SeedWordsMarkup/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SeedWordsMarkup';
