+++ src/containers/SideBar/components/Settings/Markups/SeedWordsMarkup/useGetSeedWords.ts
