+++ src/containers/SideBar/components/Settings/Settings.tsx
@@ -1,20 +1,9 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 
-import {
-    IoSettingsOutline,
-    IoCopyOutline,
-    IoEyeOutline,
-    IoEyeOffOutline,
-    IoCheckmarkOutline,
-    IoClose,
-} from 'react-icons/io5';
-import { useGetSeedWords } from '../../../../hooks/useGetSeedWords';
+import { IoSettingsOutline, IoClose } from 'react-icons/io5';
 
-import { invoke } from '@tauri-apps/api/tauri';
-
-import { useAppStatusStore } from '@app/store/useAppStatusStore.ts';
 import VisualMode from '../../../Dashboard/components/VisualMode';
-import { CardContainer, HorisontalBox, HeadingContainer } from './Settings.styles';
+import { HorisontalBox, HeadingContainer } from './Settings.styles';
 
 import AirdropPermissionSettings from '@app/containers/Airdrop/AirdropPermissionSettings/AirdropPermissionSettings.tsx';
 import LogsSettings from './LogsSettings';
@@ -26,378 +15,72 @@ import { Typography } from '@app/components/elements/Typography.tsx';
 import { Divider } from '@app/components/elements/Divider.tsx';
 import { SettingsTabs } from '@app/components/elements/Tabs';
 
-import { CircularProgress } from '@app/components/elements/CircularProgress.tsx';
 import AppVersions from '@app/containers/SideBar/components/Settings/AppVersions.tsx';
 import LanguageSettings from '@app/containers/SideBar/components/Settings/LanguageSettings.tsx';
 import HardwareStatus from '@app/containers/SideBar/components/Settings/HardwareStatus.tsx';
 
 import DebugSettings from '@app/containers/SideBar/components/Settings/DebugSettings.tsx';
-import { MinerContainer } from '../../Miner/styles';
 import { useTranslation } from 'react-i18next';
-import { ToggleSwitch } from '@app/components/elements/ToggleSwitch.tsx';
-import { useAppStateStore } from '@app/store/appStateStore.ts';
-import { useCPUStatusStore } from '@app/store/useCPUStatusStore.ts';
-import { useShallow } from 'zustand/react/shallow';
 
 import { ResetSettingsButton } from '@app/containers/SideBar/components/Settings/ResetSettingsButton.tsx';
-import { useMiningStore } from '@app/store/useMiningStore.ts';
-import { useGPUStatusStore } from '@app/store/useGPUStatusStore.ts';
-import { SeedWords } from './SeedWords';
-import { CardComponent } from '@app/containers/SideBar/components/Settings/Card.component.tsx';
-import MoneroAddressEditor from './MoneroAddressEditor';
+import MoneroAddressMarkup from './Markups/MoneroAddressMarkup';
+import WalletAddressMarkup from './Markups/WalletAddressMarkup';
+import CpuMiningMarkup from './Markups/CpuMiningMarkup';
+import P2pMarkup from './Markups/P2pMarkup';
+import GpuMiningMarkup from './Markups/GpuMiningMarkup';
+import SeedWordsMarkup from './Markups/SeedWordsMarkup';
+
+const GeneralTab = () => (
+    <Stack gap={10}>
+        <MoneroAddressMarkup />
+        <Divider />
+        <SeedWordsMarkup />
+        <Divider />
+        <HorisontalBox>
+            <CpuMiningMarkup />
+            <GpuMiningMarkup />
+        </HorisontalBox>
+        <Divider />
+        <LogsSettings />
+        <Divider />
+        <LanguageSettings />
+        <Divider />
+        <AirdropPermissionSettings />
+        <Divider />
+        <HorisontalBox>
+            <ResetSettingsButton />
+        </HorisontalBox>
+    </Stack>
+);
+
+const ExperimentalTab = () => (
+    <Stack gap={10}>
+        <WalletAddressMarkup />
+        <Divider />
+        <P2pMarkup />
+        <Divider />
+        <DebugSettings />
+        <Divider />
+        <P2pMarkup />
+        <Divider />
+        <HardwareStatus />
+        <Divider />
+        <AppVersions />
+        <Divider />
+        <Stack direction="row" justifyContent="space-between">
+            <VisualMode />
+        </Stack>
+    </Stack>
+);
+
+const tabs = [
+    { label: 'General', content: <GeneralTab /> },
+    { label: 'Experimental', content: <ExperimentalTab /> },
+];
 
 export default function Settings() {
     const { t } = useTranslation(['common', 'settings'], { useSuspense: false });
-
-    const {
-        moneroAddress,
-        walletAddress,
-        isCpuMiningEnabled,
-        isGpuMiningEnabled,
-        isP2poolEnabled,
-        p2poolStats,
-        walletAddressEmoji,
-        setMoneroAddress,
-    } = useAppStatusStore(
-        useShallow((s) => ({
-            moneroAddress: s.monero_address,
-            walletAddress: s.tari_address_base58,
-            walletAddressEmoji: s.tari_address_emoji,
-            isCpuMiningEnabled: s.cpu_mining_enabled,
-            isGpuMiningEnabled: s.gpu_mining_enabled,
-            isP2poolEnabled: s.p2pool_enabled,
-            p2poolStats: s.p2pool_stats,
-            setMoneroAddress: s.setMoneroAddress,
-        }))
-    );
-
-    // p2pool
-    const p2poolSha3Stats = p2poolStats?.sha3;
-    const p2poolRandomXStats = p2poolStats?.randomx;
-    const p2poolTribe = p2poolSha3Stats?.tribe?.name;
-    const p2poolSha3MinersCount = p2poolSha3Stats?.num_of_miners;
-    const p2poolRandomxMinersCount = p2poolRandomXStats?.num_of_miners;
-    const p2poolSha3HashRate = p2poolSha3Stats?.pool_hash_rate;
-    const p2poolRandomxHashRate = p2poolRandomXStats?.pool_hash_rate;
-    const p2poolSha3TotalEarnings = p2poolSha3Stats?.pool_total_earnings;
-    const p2poolRandomxTotalEarnings = p2poolRandomXStats?.pool_total_earnings;
-    const p2poolSha3ChainTip = p2poolSha3Stats?.share_chain_height;
-    const p2poolRandomxChainTip = p2poolRandomXStats?.share_chain_height;
-    const p2poolSha3UserTotalEarnings = walletAddress ? p2poolSha3Stats?.total_earnings[walletAddress] : 0;
-    const p2poolRandomxUserTotalEarnings = walletAddress ? p2poolRandomXStats?.total_earnings[walletAddress] : 0;
-    const p2poolUserTotalEarnings =
-        p2poolSha3UserTotalEarnings && p2poolRandomxUserTotalEarnings
-            ? p2poolSha3UserTotalEarnings + p2poolRandomxUserTotalEarnings
-            : 0;
-
-    const [showSeedWords, setShowSeedWords] = useState(false);
-    const [isCopyTooltipHidden, setIsCopyTooltipHidden] = useState(true);
-    const [isCopyTooltipHiddenWalletAddress, setIsCopyTooltipHiddenWalletAddress] = useState(true);
-    const { seedWords, getSeedWords, seedWordsFetched, seedWordsFetching } = useGetSeedWords();
-    const miningAllowed = useAppStateStore(useShallow((s) => s.setupProgress >= 1));
-    const setError = useAppStateStore((s) => s.setError);
-    const isCPUMining = useCPUStatusStore(useShallow((s) => s.is_mining));
-    const isGPUMining = useGPUStatusStore(useShallow((s) => s.is_mining));
-    const isMiningInProgress = isCPUMining || isGPUMining;
-    const miningInitiated = useMiningStore(useShallow((s) => s.miningInitiated));
-    const miningLoading = (miningInitiated && !isMiningInProgress) || (!miningInitiated && isMiningInProgress);
     const [open, setOpen] = useState(false);
-
-    const toggleSeedWordsVisibility = async () => {
-        if (!seedWordsFetched) {
-            await getSeedWords();
-        }
-        setShowSeedWords((p) => !p);
-    };
-
-    const copySeedWords = async () => {
-        if (!seedWordsFetched) {
-            await getSeedWords();
-        }
-        setIsCopyTooltipHidden(false);
-        await navigator.clipboard.writeText(seedWords.join(' '));
-        setTimeout(() => setIsCopyTooltipHidden(true), 1000);
-    };
-
-    const copyWalletAddress = async () => {
-        setIsCopyTooltipHiddenWalletAddress(false);
-        await navigator.clipboard.writeText(walletAddress + '');
-        setTimeout(() => setIsCopyTooltipHiddenWalletAddress(true), 1000);
-    };
-
-    const handleMoneroAddressChange = async (moneroAddress: string) => {
-        try {
-            console.info('Setting Monero address', moneroAddress);
-            await invoke('set_monero_address', { moneroAddress });
-            setMoneroAddress(moneroAddress);
-        } catch (error) {
-            console.error('Failed to set Monero address', error);
-            setError('Failed to set Monero address!');
-        }
-    };
-
-    useEffect(() => {
-        const fetchInitialData = async () => {
-            const fetchedMoneroAddress = await invoke('get_monero_address');
-            setMoneroAddress(fetchedMoneroAddress);
-        };
-
-        fetchInitialData();
-    }, [setMoneroAddress]);
-
-    const walletAddressMarkup = walletAddress ? (
-        <>
-            <Stack>
-                <Stack direction="row" justifyContent="space-between" style={{ height: 40 }}>
-                    <Typography variant="h6">Tari Wallet Address</Typography>
-                </Stack>
-                <Stack direction="row" justifyContent="space-between">
-                    <Typography variant="p">{walletAddress}</Typography>
-                    <IconButton onClick={copyWalletAddress}>
-                        {isCopyTooltipHiddenWalletAddress ? <IoCopyOutline /> : <IoCheckmarkOutline />}
-                    </IconButton>
-                </Stack>
-                <Stack direction="row" justifyContent="space-between">
-                    <Typography variant="p">{walletAddressEmoji}</Typography>
-                </Stack>
-            </Stack>
-        </>
-    ) : null;
-
-    const seedWordMarkup = (
-        <Stack>
-            <Stack direction="row" justifyContent="space-between" style={{ height: 40 }}>
-                <Typography variant="h6">Seed Words</Typography>
-                {showSeedWords && seedWordsFetched && (
-                    <IconButton onClick={copySeedWords}>
-                        {isCopyTooltipHidden ? <IoCopyOutline /> : <IoCheckmarkOutline />}
-                    </IconButton>
-                )}
-            </Stack>
-            <Stack direction="row" justifyContent="space-between">
-                <SeedWords showSeedWords={showSeedWords} seedWords={seedWords} />
-                <IconButton onClick={toggleSeedWordsVisibility} disabled={seedWordsFetching}>
-                    {seedWordsFetching ? (
-                        <CircularProgress />
-                    ) : showSeedWords ? (
-                        <IoEyeOffOutline size={18} />
-                    ) : (
-                        <IoEyeOutline size={18} />
-                    )}
-                </IconButton>
-            </Stack>
-        </Stack>
-    );
-
-    const moneroAddressMarkup = (
-        <Stack>
-            <Stack direction="row" justifyContent="space-between" style={{ height: 40 }}>
-                <Typography variant="h6">Monero Address</Typography>
-            </Stack>
-            <Stack direction="row" justifyContent="space-between">
-                <MoneroAddressEditor initialAddress={moneroAddress || ''} onApply={handleMoneroAddressChange} />
-            </Stack>
-        </Stack>
-    );
-    const handleP2poolEnabled = (event: React.ChangeEvent<HTMLInputElement>) => {
-        const isChecked = event.target.checked;
-        invoke('set_p2pool_enabled', { p2poolEnabled: isChecked }).then(() => {
-            console.info('P2pool enabled checked', isChecked);
-        });
-    };
-
-    const handleCpuMiningEnabled = useCallback(async () => {
-        await invoke('set_cpu_mining_enabled', { enabled: !isCpuMiningEnabled });
-    }, [isCpuMiningEnabled]);
-
-    const handleGpuMiningEnabled = useCallback(async () => {
-        await invoke('set_gpu_mining_enabled', { enabled: !isGpuMiningEnabled });
-    }, [isGpuMiningEnabled]);
-
-    const p2pMarkup = (
-        <MinerContainer>
-            <Stack>
-                <Typography variant="h6">{t('pool-mining', { ns: 'settings' })}</Typography>
-                <Stack direction="row" justifyContent="space-between">
-                    <Typography variant="p">{t('pool-mining-description', { ns: 'settings' })}</Typography>
-                    <ToggleSwitch
-                        checked={isP2poolEnabled}
-                        disabled={isMiningInProgress || !miningAllowed || miningLoading}
-                        onChange={handleP2poolEnabled}
-                    />
-                </Stack>
-            </Stack>
-        </MinerContainer>
-    );
-
-    const toggleDisabledBase = !miningAllowed || miningLoading;
-    const cpuDisabled = isMiningInProgress && isCpuMiningEnabled && !isGpuMiningEnabled; // TODO: should we rather stop mining if they both get turned off from settings?
-    const gpuDisabled = isMiningInProgress && isGpuMiningEnabled && !isCpuMiningEnabled;
-
-    const cpuEnabledMarkup = (
-        <MinerContainer>
-            <Typography variant="h6">{t('cpu-mining-enabled', { ns: 'settings' })}</Typography>
-            <ToggleSwitch
-                checked={isCpuMiningEnabled}
-                disabled={toggleDisabledBase || cpuDisabled}
-                onChange={handleCpuMiningEnabled}
-            />
-        </MinerContainer>
-    );
-
-    const gpuEnabledMarkup = (
-        <MinerContainer>
-            <Typography variant="h6">{t('gpu-mining-enabled', { ns: 'settings' })}</Typography>
-            <ToggleSwitch
-                checked={isGpuMiningEnabled}
-                disabled={toggleDisabledBase || gpuDisabled}
-                onChange={handleGpuMiningEnabled}
-            />
-        </MinerContainer>
-    );
-
-    const p2poolStatsMarkup = (
-        <>
-            <MinerContainer>
-                <HorisontalBox>
-                    <Typography variant="h6">{t('p2pool-stats', { ns: 'settings' })}</Typography>
-                </HorisontalBox>
-                <CardContainer>
-                    <CardComponent
-                        heading={`${t('tribe', { ns: 'settings' })}`}
-                        labels={[
-                            {
-                                labelText: 'Current',
-                                labelValue: p2poolTribe ? p2poolTribe : '',
-                            },
-                        ]}
-                    />
-                    <CardComponent
-                        heading={`${t('miners', { ns: 'settings' })}`}
-                        labels={[
-                            {
-                                labelText: 'SHA-3',
-                                labelValue: '' + p2poolSha3MinersCount,
-                            },
-                            {
-                                labelText: 'RandomX',
-                                labelValue: '' + p2poolRandomxMinersCount,
-                            },
-                        ]}
-                    />
-                    <CardComponent
-                        heading={`${t('p2pool-hash-rate', { ns: 'settings' })}`}
-                        labels={[
-                            {
-                                labelText: 'SHA-3',
-                                labelValue: (p2poolSha3HashRate ? p2poolSha3HashRate : 0) + ' H/s',
-                            },
-                            {
-                                labelText: 'RandomX',
-                                labelValue: (p2poolRandomxHashRate ? p2poolRandomxHashRate : 0) + ' H/s',
-                            },
-                        ]}
-                    />
-                    <CardComponent
-                        heading={`${t('p2pool-total-earnings', { ns: 'settings' })}`}
-                        labels={[
-                            {
-                                labelText: 'SHA-3',
-                                labelValue: (p2poolSha3TotalEarnings ? p2poolSha3TotalEarnings : 0) + ' tXTM',
-                            },
-                            {
-                                labelText: 'RandomX',
-                                labelValue: (p2poolRandomxTotalEarnings ? p2poolRandomxTotalEarnings : 0) + ' tXTM',
-                            },
-                        ]}
-                    />
-                    <CardComponent
-                        heading={`${t('p2pool-chain-tip', { ns: 'settings' })}`}
-                        labels={[
-                            {
-                                labelText: 'SHA-3',
-                                labelValue: '#' + p2poolSha3ChainTip,
-                            },
-                            {
-                                labelText: 'RandomX',
-                                labelValue: '#' + p2poolRandomxChainTip,
-                            },
-                        ]}
-                    />
-                    <CardComponent
-                        heading={`${t('p2pool-user-total-earnings', { ns: 'settings' })}`}
-                        labels={[
-                            {
-                                labelText: 'SHA-3',
-                                labelValue: (p2poolSha3UserTotalEarnings ? p2poolSha3UserTotalEarnings : 0) + ' tXTM',
-                            },
-                            {
-                                labelText: 'RandomX',
-                                labelValue:
-                                    (p2poolRandomxUserTotalEarnings ? p2poolRandomxUserTotalEarnings : 0) + ' tXTM',
-                            },
-                            {
-                                labelText: 'Total',
-                                labelValue: p2poolUserTotalEarnings + ' tXTM',
-                            },
-                        ]}
-                    />
-                </CardContainer>
-            </MinerContainer>
-        </>
-    );
-
-    const GeneralTab = () => {
-        return (
-            <Stack gap={10}>
-                {moneroAddressMarkup}
-                <Divider />
-                {seedWordMarkup}
-                <Divider />
-                <HorisontalBox>
-                    {cpuEnabledMarkup}
-                    {gpuEnabledMarkup}
-                </HorisontalBox>
-                <Divider />
-                <LogsSettings />
-                <Divider />
-                <LanguageSettings />
-                <Divider />
-                <AirdropPermissionSettings />
-                <Divider />
-                <HorisontalBox>
-                    <ResetSettingsButton />
-                </HorisontalBox>
-            </Stack>
-        );
-    };
-
-    const ExperimentalTab = () => {
-        return (
-            <Stack gap={10}>
-                {walletAddressMarkup}
-                <Divider />
-                {p2pMarkup}
-                <Divider />
-                <DebugSettings />
-                <Divider />
-                {p2poolStatsMarkup}
-                <Divider />
-                <HardwareStatus />
-                <Divider />
-                <AppVersions />
-                <Divider />
-                <Stack direction="row" justifyContent="space-between">
-                    <VisualMode />
-                </Stack>
-            </Stack>
-        );
-    };
-
-    const tabs = [
-        { label: 'General', content: <GeneralTab /> },
-        { label: 'Experimental', content: <ExperimentalTab /> },
-    ];
 
     return (
         <Dialog open={open} onOpenChange={setOpen}>
